fix: tech 패키지 및 BaseTimeEntity 생성
