+++ src/main/java/com/potential_radar/PR/common/entity/BaseTimeEntity.java
@@ -1,0 +1,18 @@
+package com.potential_radar.PR.common.entity;
+
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass
+public class BaseTimeEntity {
+    @CreationTimestamp
+    private LocalDateTime createdTime;
+
+    @UpdateTimestamp
+    private LocalDateTime updatedTime;
+}
