+++ src/main/java/com/potential_radar/PR/project/domain/ProjectTechPart.java
@@ -1,7 +1,6 @@
 package com.potential_radar.PR.project.domain;
 
-import com.potential_radar.PR.common.entity.TechPart;
-import com.potential_radar.PR.common.entity.TechStack;
+import com.potential_radar.PR.tech.entity.TechPart;
 import jakarta.persistence.*;
 import lombok.*;
