+++ src/main/java/com/potential_radar/PR/project/domain/ProjectTechStack.java
@@ -1,6 +1,6 @@
 package com.potential_radar.PR.project.domain;
 
-import com.potential_radar.PR.common.entity.TechStack;
+import com.potential_radar.PR.tech.entity.TechStack;
 import jakarta.persistence.*;
 import lombok.*;
