+++ src/main/java/com/potential_radar/PR/tech/repository/TechPartRepository.java
@@ -1,6 +1,6 @@
-package com.potential_radar.PR.common.repository;
+package com.potential_radar.PR.tech.repository;
 
-import com.potential_radar.PR.common.entity.TechPart;
+import com.potential_radar.PR.tech.entity.TechPart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
